[level 2] Title: k진수에서 소수 개수 구하기, Time: 1.33 ms, Memory: 75 MB -BaekjoonHub

+++ 프로그래머스/lv2/92335. k진수에서 소수 개수 구하기/k진수에서 소수 개수 구하기.java
@@ -8,7 +8,7 @@ class Solution {
             if (c != '0') {
                 temp += String.valueOf(c);
             }else {
-                if (!temp.equals("") && isPrime(Integer.parseInt(temp))) {
+                if (!temp.equals("") && isPrime(Long.parseLong(temp))) {
                     answer ++;
                 }
                 temp = "";
